Date param added to external endpoinsta calls

+++ Server/model/load_data_job.go
@@ -4,6 +4,6 @@ package model
 type LoadDataJob struct {
 	UID   string `predicate:"uid,omitempty"`
 	ID    string `predicate:"load_data_job_id,omitempty"`
-	Date  int    `predicate:"load_data_job_date,omitempty"`
+	Date  int64  `predicate:"load_data_job_date,omitempty"`
 	Email string `predicate:"load_data_job_email,omitempty"`
 }
